pyproject.toml
[project]
name = "MakerHub"
version = "0.1.0"
description = "MakerHub"
authors = [
    { name = "assumeengage", email = "xsk13969662263@outlook.com" }
]
license = { text = "MIT" }
readme = "README.md"
package-mode = false

[tool.poetry]
name = "MakerHub"
version = "0.1.0"
description = "MakerHub"
authors = ["assumeengage"]
license = "MIT"
readme = "false"


[tool.poetry.dependencies]
python = "^3.9"
pyjwt = "^2.0"
passlib = "^1.7"
fastapi = "^0.95.0"
uvicorn = "^0.15.0"
pymongo = "^3.12.0"
minio = "^7.1.0"
motor = "^2.5"
loguru = "^0.5.3"
pydantic = "^1.8.2"
aiohttp = "^3.8.1"
python-dotenv = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"